Remove tempusdominus dependency
Instead we use native input of type date, datetime-local and time to pop-up a
calendar. Such input can not be directly used because we can not control the
format, so we synchronize the value with a text input using the proper format.
The text input is also parsed using the right format.

issue9922
review339181002