Do not try to expand pseudo-children of reference field
Since issue5667 the export dialog adds pseudo children to reference fields but
they are not children nodes but just sibling. So we must not try to expand the
node of the reference field when loading an predefined export.

issue10875
review383511003
(grafted from 0fb9992f455702540f0aa4ef82ec3260a06907a1)